color codes time blocks according to current time
New function in JS compares the current time to the integer value of the time in the individual time blocks and displays them accordingly in relation to the current time.
Minor changes had to be made to the html to get this function to work properly. Changed integer value of the times in the html to work with moment.js times.